+++ packages/node-core/src/indexer/blockDispatcher/worker-block-dispatcher.spec.ts
@@ -4,7 +4,7 @@
 import {EventEmitter2} from '@nestjs/event-emitter';
 import {IProjectUpgradeService, NodeConfig} from '../../configure';
 import {DynamicDsService} from '../dynamic-ds.service';
-import {PoiService, PoiSyncService} from '../poi';
+import {PoiSyncService} from '../poi';
 import {SmartBatchService} from '../smartBatch.service';
 import {StoreService} from '../store.service';
 import {StoreCacheService} from '../storeCache';
@@ -35,7 +35,6 @@ describe('WorkerBlockDispatcher', () => {
       {minimumHeapRequired: 150} as unknown as SmartBatchService,
       null as unknown as StoreService,
       null as unknown as StoreCacheService,
-      null as unknown as PoiService,
       null as unknown as PoiSyncService,
       null as unknown as ISubqueryProject,
       null as unknown as DynamicDsService<any>,