+++ packages/node-core/src/indexer/unfinalizedBlocks.service.ts
@@ -2,15 +2,19 @@
 // SPDX-License-Identifier: GPL-3.0
 
 import assert from 'assert';
-import {last} from 'lodash';
+import {isEqual, last} from 'lodash';
 import {NodeConfig} from '../configure';
 import {getLogger} from '../logger';
+import {ProofOfIndex} from './entities';
+import {PoiBlock} from './poi';
 import {StoreCacheService} from './storeCache';
 
 const logger = getLogger('UnfinalizedBlocks');
 
 export const METADATA_UNFINALIZED_BLOCKS_KEY = 'unfinalizedBlocks';
 export const METADATA_LAST_FINALIZED_PROCESSED_KEY = 'lastFinalizedVerifiedHeight';
+
+export const POI_NOT_ENABLED_ERROR_MESSAGE = 'Poi is not enabled, unable to check for last finalized block';
 
 const UNFINALIZED_THRESHOLD = 200;
 
@@ -67,15 +71,14 @@ export abstract class BaseUnfinalizedBlocksService<B> implements IUnfinalizedBlo
     this.lastCheckedBlockHeight = await this.getLastFinalizedVerifiedHeight();
     this.finalizedHeader = await this.getFinalizedHead();
 
-    if (!this.nodeConfig.unfinalizedBlocks && this.unfinalizedBlocks.length) {
+    if (this.unfinalizedBlocks.length) {
       logger.info('Processing unfinalized blocks');
       // Validate any previously unfinalized blocks
 
       const rewindHeight = await this.processUnfinalizedBlocks();
-
       if (rewindHeight !== undefined) {
         logger.info(
-          `Found un-finalized blocks from previous indexing but unverified, rolling back to last finalized block ${rewindHeight} `
+          `Found un-finalized blocks from previous indexing but unverified, rolling back to last finalized block ${rewindHeight}`
         );
         await reindex(rewindHeight);
         logger.info(`Successful rewind to block ${rewindHeight}!`);
@@ -153,13 +156,9 @@ export abstract class BaseUnfinalizedBlocksService<B> implements IUnfinalizedBlo
   // find closest record from block heights
   private getClosestRecord(blockHeight: number): Header | undefined {
     // Have the block in the best block, can be verified
-    const record = [...this.unfinalizedBlocks] // Copy so we can reverse
+    return [...this.unfinalizedBlocks] // Copy so we can reverse
       .reverse() // Reverse the list to find the largest block
       .find(({blockHeight: height}) => height <= blockHeight);
-    if (record) {
-      return record;
-    }
-    return undefined;
   }
 
   // check unfinalized blocks for a fork, returns the header where a fork happened
@@ -175,7 +174,7 @@ export abstract class BaseUnfinalizedBlocksService<B> implements IUnfinalizedBlo
     if (lastVerifiableBlock.blockHeight === this.finalizedBlockNumber) {
       if (lastVerifiableBlock.blockHash !== this.finalizedHeader.blockHash) {
         logger.warn(
-          `Block fork found, enqueued un-finalized block at ${lastVerifiableBlock.blockHeight} with hash ${lastVerifiableBlock.blockHash}, actual hash is ${this.finalizedHeader.blockHash}`
+          `Block fork found, enqueued un-finalized block at ${lastVerifiableBlock.blockHeight} with hash ${lastVerifiableBlock.blockHash}, actual hash is ${this.finalizedHeader.blockHash}.`
         );
         return this.finalizedHeader;
       }
@@ -224,6 +223,48 @@ export abstract class BaseUnfinalizedBlocksService<B> implements IUnfinalizedBlo
     }
 
     return this.lastCheckedBlockHeight;
+  }
+
+  // Finds the last POI that had a correct block hash, this is used with the Eth sdk
+  protected async findFinalizedUsingPOI(header: Header): Promise<Header> {
+    const poiModel = this.storeCache.poi;
+    if (!poiModel) {
+      throw new Error(POI_NOT_ENABLED_ERROR_MESSAGE);
+    }
+
+    let lastHeight = header.blockHeight;
+
+    // eslint-disable-next-line no-constant-condition
+    while (true) {
+      const indexedBlocks: ProofOfIndex[] = await poiModel.getPoiBlocksBefore(lastHeight);
+
+      if (!indexedBlocks.length) {
+        break;
+      }
+
+      // Work backwards to find a block on chain that matches POI
+      for (const indexedBlock of indexedBlocks) {
+        const chainHeader = await this.getHeaderForHeight(indexedBlock.id);
+
+        // Need to convert to PoiBlock to encode block hash to Uint8Array properly
+        const testPoiBlock = PoiBlock.create(
+          chainHeader.blockHeight,
+          chainHeader.blockHash,
+          new Uint8Array(),
+          indexedBlock.projectId ?? ''
+        );
+
+        // Need isEqual because of Uint8Array type
+        if (isEqual(testPoiBlock.chainBlockHash, indexedBlock.chainBlockHash)) {
+          return chainHeader;
+        }
+      }
+
+      // Next page of POI, use height rather than offset/limit as data could change in that time
+      lastHeight = indexedBlocks[indexedBlocks.length - 1].id - 1;
+    }
+
+    throw new Error('Unable to find a POI block with matching block hash');
   }
 
   private saveUnfinalizedBlocks(unfinalizedBlocks: UnfinalizedBlocks): void {